+++ backend/src/main/java/org/petzonalize/backend/repository/CustomizableRepository.java
@@ -1,8 +1,8 @@
 package org.petzonalize.backend.repository;
 
-import org.petzonalize.backend.entity.model.Customizable;
+import org.petzonalize.backend.entity.Customizable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomizableRepository extends JpaRepository<Customizable, Integer> {
+public interface CustomizableRepository extends JpaRepository<Customizable, Long> {
 
 }
