+++ backend/src/main/java/org/petzonalize/backend/repository/OrderHasProductRepository.java
@@ -1,8 +1,8 @@
 package org.petzonalize.backend.repository;
 
-import org.petzonalize.backend.entity.model.OrderHasProduct;
+import org.petzonalize.backend.entity.OrderHasProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderHasProductRepository extends JpaRepository<OrderHasProduct, Integer> {
+public interface OrderHasProductRepository extends JpaRepository<OrderHasProduct, Long> {
 
 }
