+++ backend/src/main/java/org/petzonalize/backend/repository/OrderRepository.java
@@ -1,8 +1,8 @@
 package org.petzonalize.backend.repository;
 
-import org.petzonalize.backend.entity.model.Order;
+import org.petzonalize.backend.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 
 }
