+++ backend/src/main/java/org/petzonalize/backend/repository/ProductRepository.java
@@ -2,9 +2,9 @@ package org.petzonalize.backend.repository;
 
 import java.util.Optional;
 
-import org.petzonalize.backend.entity.model.Product;
+import org.petzonalize.backend.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 	Optional<Product> findByName(String name);
 }
