+++ backend/src/main/java/org/petzonalize/backend/repository/UserRepository.java
@@ -2,10 +2,10 @@ package org.petzonalize.backend.repository;
 
 import java.util.Optional;
 
-import org.petzonalize.backend.entity.model.User;
+import org.petzonalize.backend.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 	Optional<User> findByEmail(String email);
 	void deleteByEmail(String email);
 }
