+++ backend/src/main/java/org/petzonalize/backend/service/ProductService.java
@@ -1,11 +1,11 @@
 package org.petzonalize.backend.service;
 
-import org.petzonalize.backend.entity.model.Product;
+import org.petzonalize.backend.entity.Product;
 import org.springframework.http.ResponseEntity;
 
 public interface ProductService {
     ResponseEntity<?> getProducts();
 	ResponseEntity<?> createProduct(Product product);
-	ResponseEntity<String> deleteProduct(int id);
+	ResponseEntity<String> deleteProduct(Long id);
 	ResponseEntity<?> updateProduct(Product product);
 }
